+++ src/lib.rs
@@ -48,6 +48,7 @@
 #[cfg(feature = "print-file")]
 use std::path::Path;
 
+use console::Term;
 use crossterm::{
     cursor::{RestorePosition, SavePosition},
     execute,
@@ -97,7 +98,9 @@ pub fn print(img: &DynamicImage, config: &Config) -> ViuResult<(u32, u32)> {
         execute!(&mut stdout, SavePosition)?;
     }
 
-    let (w, h) = choose_printer(config).print(&mut stdout, img, config)?;
+    let term = Term::stdout();
+
+    let (w, h) = choose_printer(config).print(&mut stdout, &term, img, config)?;
 
     if config.restore_cursor {
         execute!(&mut stdout, RestorePosition)?;
@@ -128,7 +131,9 @@ pub fn print_from_file<P: AsRef<Path>>(filename: P, config: &Config) -> ViuResul
         execute!(&mut stdout, SavePosition)?;
     }
 
-    let (w, h) = choose_printer(config).print_from_file(&mut stdout, filename, config)?;
+    let term = Term::stdout();
+
+    let (w, h) = choose_printer(config).print_from_file(&mut stdout, &term, filename, config)?;
 
     if config.restore_cursor {
         execute!(&mut stdout, RestorePosition)?;